+++ src/main/java/com/box/sdk/BoxAPIConnection.java
@@ -36,7 +36,7 @@ public class BoxAPIConnection {
     private static final String BOX_NOTIFICATIONS_HEADER = "Box-Notifications";
 
     private static final String JAVA_VERSION = System.getProperty("java.version");
-    private static final String SDK_VERSION = "2.19.0";
+    private static final String SDK_VERSION = "2.20.0";
 
     /**
      * The amount of buffer time, in milliseconds, to use when determining if an access token should be refreshed. For