+++ src/main/java/com/box/sdk/Metadata.java
@@ -226,6 +226,22 @@ public class Metadata {
     }
 
     /**
+     * Tests that a list of properties has the expected value.
+     * The values passed in will have to be an exact match with no extra elements.
+     * @param path      the path that designates the key. Must be prefixed with a "/".
+     * @param values    the list of expected values.
+     * @return          this metadata object.
+     */
+    public Metadata test(String path, List<String> values) {
+        JsonArray arr = new JsonArray();
+        for (String value : values) {
+            arr.add(value);
+        }
+        this.addOp("test", path, arr);
+        return this;
+    }
+
+    /**
      * Returns a value.
      * @param path the path that designates the key. Must be prefixed with a "/".
      * @return the metadata property value.